RSA for register works, keyring in progress

+++ register.py
@@ -4,12 +4,14 @@ import json
 import os
 import sys
 
+import gnupg
 from PyQt5 import QtWidgets, uic
 from PyQt5.QtCore import Qt
 from PyQt5.QtWidgets import QMessageBox
 
 qt_creator_file = "guis/register.ui"
 Ui_MainWindow, QtBaseClass = uic.loadUiType(qt_creator_file)
+gpg = gnupg.GPG(gnupghome="/home/marina/gnupg")
 
 
 def hash_password(password, salt):
@@ -54,11 +56,28 @@ class MainWindow(QtWidgets.QMainWindow, Ui_MainWindow):
         if not email or not master_password:
             QMessageBox.about(self, "No data", "Write password name and password, please")
         else:
-            hashed = hash_password(master_password, salt)
-            with open('register.json', 'w+') as file:
-                data = {'email': email, 'master_password': hashed, 'salt': salt.decode()}
-                json.dump(data, file)
+            if os.path.exists('register.json.gpg'):
+                self.show_message_box()
+            else:
+                hashed = hash_password(master_password, salt)
+                with open('register.json', 'x') as file:
+                    data = {'email': email, 'master_password': hashed, 'salt': salt.decode(),
+                            'directory': self.directory.text()}
+                    json.dump(data, file)
+                input_data = gpg.gen_key_input(
+                    name_email='marynasavchen@gmail.com',
+                    passphrase='passphrase')
+                gpg.gen_key(input_data)
+                with open('register.json', 'rb') as file:
+                    gpg.encrypt_file(file, recipients=['marynasavchen@gmail.com'], output='register.json.gpg')
+                os.remove('register.json')
+
             self.on_cancel_button()
+
+    def show_message_box(self):
+        """Show MessageBox with error if there is no such user. Clear fields"""
+        QMessageBox.about(self, "One account", "Only one account is possible")
+        self.on_cancel_button()
 
 
 if __name__ == '__main__':
